% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_wen_sthd_data.R
\name{prep_wen_sthd_data}
\alias{prep_wen_sthd_data}
\title{Prepare All Data from Wenatchee for a Report}
\usage{
prep_wen_sthd_data(
  redd_file_path =
    "T:/DFW-Team FP Upper Columbia Escapement - General/UC_Sthd/inputs/Redd Data",
  redd_file_name = "Wenatchee_Redd_Surveys.xlsx",
  dabom_file_path =
    "T:/DFW-Team FP Upper Columbia Escapement - General/UC_Sthd/estimates",
  dabom_file_name = "UC_STHD_Model_Output.xlsx",
  n_observers = "two",
  query_year = lubridate::year(lubridate::today()) - 1
)
}
\value{
dataframe
}
\description{
This function
}
\author{
Kevin See
}
