% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_redd_data.R
\name{query_redd_data}
\alias{query_redd_data}
\title{Query Redd Data}
\usage{
query_redd_data(
  file_path =
    "T:/DFW-Team FP Upper Columbia Escapement - General/UC_Sthd/inputs/Redd Data",
  file_name = "Wenatchee_Redd_Surveys.xlsx",
  query_year = lubridate::year(lubridate::today()) - 1
)
}
\arguments{
\item{file_path}{path to data file}

\item{file_name}{name of Excel file containing redd data in a very particular format}

\item{query_year}{which year or years should be included in this query?}
}
\value{
dataframe
}
\description{
Prepares data by querying specific file on a WDFW Sharepoint site, and filters for selected year(s).
}
\author{
Kevin See
}
