% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SthdReddsWenatchee-package.R
\docType{package}
\name{SthdReddsWenatchee-package}
\alias{SthdReddsWenatchee}
\alias{SthdReddsWenatchee-package}
\title{SthdReddsWenatchee: Estimate Steelhead Redds in Wenatchee Subbasin}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wdfw-fp/SthdReddsWenatchee}
  \item Report bugs at \url{https://github.com/wdfw-fp/SthdReddsWenatchee/issues}
}

}
\author{
\strong{Maintainer}: Kevin See \email{Kevin.See@dfw.wa.gov}

}
\keyword{internal}
